# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Distribuidora API",
    "version": "dev"
  },
  "paths": {
...
    "/api/v1/catalogo/productos/{producto_id}/costo": {
      "get": {
        "tags": [
          "catalogo"
        ],
        "summary": "Obtener Costo Promedio",
        "description": "Costo promedio vigente de un producto (change 09, enmienda a D3 del\n2026-09-30; misma idea que ADR-035 punto 5: las lecturas viven con su\nentidad). Permiso `VER_COSTOS`.\n\n`catalogo` resuelve PRIMERO el producto en la organización del token: 404 si\nno existe o es de otra organización (INV-21, SEG-07), sin revelar nada. Recién\nentonces consulta a `costeo/service.py` (dependencia permitida `catalogo ->\ncosteo`; `costeo` no importa `catalogo`). Un producto sin ingresos responde\n200 con `costo_promedio` nulo y `stock_total` 0 (D10).",
        "operationId": "obtener_costo_promedio_api_v1_catalogo_productos__producto_id__costo_get",
        "parameters": [
          {
            "name": "producto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Producto Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostoPromedioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stock/ubicaciones": {
      "post": {
        "tags": [
          "stock"
        ],
        "summary": "Crear Ubicacion",
        "operationId": "crear_ubicacion_api_v1_stock_ubicaciones_post",
        "parameters": [
          {
            "name": "Operation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Operation-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UbicacionCrearRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UbicacionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "stock"
        ],
        "summary": "Listar Ubicaciones",
        "description": "Ubicaciones de la organización, con filtro opcional por estado, paginadas\npor cursor (spec `ubicaciones`, escenario \"Listar ubicaciones\").",
        "operationId": "listar_ubicaciones_api_v1_stock_ubicaciones_get",
        "parameters": [
          {
            "name": "activo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Activo"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaUbicaciones"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stock/ubicaciones/{ubicacion_id}": {
      "put": {
        "tags": [
          "stock"
        ],
        "summary": "Modificar Ubicacion",
        "operationId": "modificar_ubicacion_api_v1_stock_ubicaciones__ubicacion_id__put",
        "parameters": [
          {
            "name": "ubicacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ubicacion Id"
            }
          },
          {
            "name": "Operation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Operation-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UbicacionModificarRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UbicacionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stock/iniciales": {
      "post": {
        "tags": [
          "stock"
        ],
        "summary": "Registrar Stock Inicial",
        "operationId": "registrar_stock_inicial_api_v1_stock_iniciales_post",
        "parameters": [
          {
            "name": "Operation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Operation-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockInicialRegistrarRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockInicialResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stock/ubicaciones/{ubicacion_id}/saldos": {
      "get": {
        "tags": [
          "stock"
        ],
        "summary": "Obtener Saldos De Ubicacion",
        "description": "Stock de una ubicación por producto en unidad base (STK-01), con las\nunidades de la presentación de referencia para mostrarlo en cajas + unidades\n(CAT-08). El promedio solo viaja con `VER_COSTOS` (D3-A). 404 si la ubicación\nes ajena o inexistente.",
        "operationId": "obtener_saldos_de_ubicacion_api_v1_stock_ubicaciones__ubicacion_id__saldos_get",
        "parameters": [
          {
            "name": "ubicacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ubicacion Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockDeUbicacionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stock/kardex": {
      "get": {
        "tags": [
          "stock"
        ],
        "summary": "Obtener Kardex",
        "description": "Kardex de un producto en una ubicación (STK-04, D11): `desde` y `hasta`\nson fechas de negocio en la zona de la organización; el acumulado es el de la\nhistoria completa. El costo de cada movimiento solo viaja con `VER_COSTOS`\n(D3-A). 404 si el producto o la ubicación son ajenos o inexistentes.",
        "operationId": "obtener_kardex_api_v1_stock_kardex_get",
        "parameters": [
          {
            "name": "producto_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Producto Id"
            }
          },
          {
            "name": "ubicacion_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ubicacion Id"
            }
          },
          {
            "name": "desde",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Desde"
            }
          },
          {
            "name": "hasta",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hasta"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KardexResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/comandos": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Sincronizar Lote",
        "description": "Procesa el lote de la cola del dispositivo (`02` §6.4, tareas 8.1 a\n8.10). `organizacion_id`, `usuario_id` y `dispositivo_id` de la sesión\nvienen siempre del token (`contexto`), nunca del cuerpo -- cada ítem\ndeclara los suyos solo para que `sync_service.procesar_lote` pueda\nrechazar una cola ajena (`ColaAjenaError`) en vez de reatribuirla en\nsilencio.",
        "operationId": "sincronizar_lote_api_v1_sync_comandos_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoteComandosRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoteComandosResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CostoPromedioResponse": {
        "properties": {
          "producto_id": {
            "type": "string",
            "format": "uuid",
            "title": "Producto Id"
          },
          "costo_promedio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Costo Promedio"
          },
          "stock_total": {
            "type": "integer",
            "title": "Stock Total"
          }
        },
        "type": "object",
        "required": [
          "producto_id",
          "costo_promedio",
          "stock_total"
        ],
        "title": "CostoPromedioResponse",
        "description": "Costo promedio vigente de un producto (CST-10) y su stock total. El\npromedio es `None` mientras el producto no tuvo ningún ingreso con costo (D10:\n\"sin costo\", no cero); viaja como string (`\"1050.000000\"`)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginaUbicaciones": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UbicacionResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "cursor_siguiente": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor Siguiente"
          }
        },
        "type": "object",
        "required": [
          "items",
          "cursor_siguiente"
        ],
        "title": "PaginaUbicaciones"
      },
...
      "StockInicialRegistrarRequest": {
        "properties": {
          "ubicacion_id": {
            "type": "string",
            "format": "uuid",
            "title": "Ubicacion Id"
          },
          "lineas": {
            "items": {
              "$ref": "#/components/schemas/LineaStockInicialRequest"
            },
            "type": "array",
            "title": "Lineas"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ubicacion_id",
          "lineas"
        ],
        "title": "StockInicialRegistrarRequest",
        "description": "`lineas` sin cota en el esquema: el dominio la valida (1 a 200,\n`LINEAS_INVALIDAS`, D5)."
      },
...
      "UbicacionCrearRequest": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "DEPOSITO",
              "VEHICULO",
              "OTRO"
            ],
            "title": "Tipo"
          },
          "requiere_toma": {
            "type": "boolean",
            "title": "Requiere Toma",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nombre",
          "tipo"
        ],
        "title": "UbicacionCrearRequest"
      },
...
      "UbicacionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "requiere_toma": {
            "type": "boolean",
            "title": "Requiere Toma"
          },
          "activo": {
            "type": "boolean",
            "title": "Activo"
          },
          "actualizado_en": {
            "type": "string",
            "format": "date-time",
            "title": "Actualizado En"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "tipo",
          "requiere_toma",
          "activo",
          "actualizado_en"
        ],
        "title": "UbicacionResponse"
      },
...
    }
  }
}